Fixing the Area Damage!
The error happened due to a discrepancy in the calculations related to the new magic stat. Shots, being without magic, caused the result to be 0 when the magic stat was considered in the calculation. This problem only impacted shots beyond shot51 due to the presence of exactly 51 different creature models, with a default magic stat of 100 being applied to shots corresponding to the creature model count. Hence, for shots exceeding this limit, the damage was inaccurately set to 0.